Man kan se alle pokemonerne Hurra! Man kan lave en pokemon, men den har ikke noget navn, beskrivelse og dens nummer er 0 :(

+++ src/main/java/com/example/pokemon/repository/PokemonRepository.java
@@ -3,6 +3,6 @@ package com.example.pokemon.repository;
 import com.example.pokemon.model.Pokemon;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PokemonRepository extends CrudRepository<Pokemon, Long> {
+public interface PokemonRepository extends CrudRepository<Pokemon, Integer>{
 
 }
